lib/agents/vector-database-integration.ts
import { BaseAgent, AgentConfig } from './base';

export interface VectorDatabaseIntegrationConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class VectorDatabaseIntegrationAgent extends BaseAgent {
  private config: VectorDatabaseIntegrationConfig;

  constructor(config: VectorDatabaseIntegrationConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Vector Database Integration implementation
    await new Promise(resolve => setTimeout(resolve, 1000));
    
    return `Vector Database Integration completed successfully for: ${input}`;
  }

  async process(data: any): Promise<any> {
    // Advanced vector database integration processing
    return { status: 'success', data };
  }
}

// FEATURE UPDATE: Vector Database Integration
// Date: 2024-02-20
import { BaseAgent, AgentConfig } from './base';

export interface VectorDatabaseIntegrationConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class VectorDatabaseIntegrationAgent extends BaseAgent {
  private config: VectorDatabaseIntegrationConfig;

  constructor(config: VectorDatabaseIntegrationConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Vector Database Integration i...
