__tests__/lib/agents/vector-database-integration.test.ts
import { VectorDatabaseIntegrationAgent } from '@/lib/agents/vector-database-integration';

describe('Vector Database IntegrationAgent', () => {
  let agent: VectorDatabaseIntegrationAgent;

  beforeEach(() => {
    agent = new VectorDatabaseIntegrationAgent({
      name: 'Test Vector Database Integration Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input successfully', async () => {
      const result = await agent.execute('test input');
      expect(result).toContain('Vector Database Integration completed successfully');
    });

    it('should handle empty input', async () => {
      const result = await agent.execute('');
      expect(result).toBeDefined();
    });
  });

  describe('process', () => {
    it('should process data correctly', async () => {
      const result = await agent.process({ test: 'data' });
      expect(result.status).toBe('success');
    });
  });
});

// FEATURE UPDATE: Vector Database Integration
// Date: 2024-02-20
import { VectorDatabaseIntegrationAgent } from '@/lib/agents/vector-database-integration';

describe('Vector Database IntegrationAgent', () => {
  let agent: VectorDatabaseIntegrationAgent;

  beforeEach(() => {
    agent = new VectorDatabaseIntegrationAgent({
      name: 'Test Vector Database Integration Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process in...
